Move 'yield_task_args' function to utils.py, rename variables for clarity
